refactor: rename, add tsyringe

+++ src/services/carService.ts
@@ -1,7 +1,10 @@
 import Car from "../models/Car";
 import { ICar } from "../interfaces/ICar";
+import { injectable } from "tsyringe";
+import createError from "http-errors";
 
-export class CarService {
+@injectable()
+class CarService {
   async registerCar(carData: ICar) {
     const car = new Car(carData);
     await car.save();
@@ -27,18 +30,39 @@ export class CarService {
   }
 
   async getCarById(id: string): Promise<ICar | null> {
-    return await Car.findById(id).exec();
+    if (!id.match(/^[0-9a-fA-F]{24}$/)) {
+      throw createError(400, "Invalid ID format");
+    }
+    const car = await Car.findById(id).exec();
+    if (!car) {
+      throw createError(404, "Car not found");
+    }
+    return car;
   }
 
   async deleteCar(id: string): Promise<ICar | null> {
-    return await Car.findByIdAndDelete(id).exec();
+    if (!id.match(/^[0-9a-fA-F]{24}$/)) {
+      throw createError(400, "Invalid ID format");
+    }
+    const car = await Car.findByIdAndDelete(id).exec();
+    if (!car) {
+      throw createError(404, "Car not found");
+    }
+    return car;
   }
 
   async updateCar(id: string, carData: Partial<ICar>): Promise<ICar | null> {
-    return await Car.findByIdAndUpdate(id, carData, {
+    if (!id.match(/^[0-9a-fA-F]{24}$/)) {
+      throw createError(400, "Invalid ID format");
+    }
+    const car = await Car.findByIdAndUpdate(id, carData, {
       new: true,
       runValidators: true,
     }).exec();
+    if (!car) {
+      throw createError(404, "Car not found");
+    }
+    return car;
   }
 
   async updateAccessory(
@@ -46,15 +70,19 @@ export class CarService {
     accessoryId: string,
     description: string,
   ): Promise<ICar | null> {
+    if (
+      !carId.match(/^[0-9a-fA-F]{24}$/) ||
+      !accessoryId.match(/^[0-9a-fA-F]{24}$/)
+    ) {
+      throw createError(400, "Invalid ID format");
+    }
     const car = await Car.findById(carId).exec();
     if (!car) {
-      return null;
+      throw createError(404, "Car not found");
     }
-
     const accessoryIndex = car.accessories.findIndex(
       (a) => a._id.toString() === accessoryId,
     );
-
     if (accessoryIndex > -1) {
       if (car.accessories[accessoryIndex].description === description) {
         car.accessories.splice(accessoryIndex, 1);
@@ -64,8 +92,9 @@ export class CarService {
     } else {
       car.accessories.push({ _id: accessoryId, description });
     }
-
     await car.save();
     return car;
   }
 }
+
+export default CarService;
